+++ sdk/go/docker/secret.go
@@ -56,28 +56,28 @@ func GetSecret(ctx *pulumi.Context,
 }
 
 // URN is this resource's unique name assigned by Pulumi.
-func (r *Secret) URN() *pulumi.URNOutput {
+func (r *Secret) URN() pulumi.URNOutput {
 	return r.s.URN()
 }
 
 // ID is this resource's unique identifier assigned by its provider.
-func (r *Secret) ID() *pulumi.IDOutput {
+func (r *Secret) ID() pulumi.IDOutput {
 	return r.s.ID()
 }
 
 // The base64 encoded data of the secret.
-func (r *Secret) Data() *pulumi.StringOutput {
-	return (*pulumi.StringOutput)(r.s.State["data"])
+func (r *Secret) Data() pulumi.StringOutput {
+	return (pulumi.StringOutput)(r.s.State["data"])
 }
 
 // User-defined key/value metadata.
-func (r *Secret) Labels() *pulumi.MapOutput {
-	return (*pulumi.MapOutput)(r.s.State["labels"])
+func (r *Secret) Labels() pulumi.MapOutput {
+	return (pulumi.MapOutput)(r.s.State["labels"])
 }
 
 // The name of the Docker secret.
-func (r *Secret) Name() *pulumi.StringOutput {
-	return (*pulumi.StringOutput)(r.s.State["name"])
+func (r *Secret) Name() pulumi.StringOutput {
+	return (pulumi.StringOutput)(r.s.State["name"])
 }
 
 // Input properties used for looking up and filtering Secret resources.